Create videodatabase.py
Here's a really simple sqlite database that uses PeeWee to create a database of all of our elements in each video asset. The reason we need a database is because (at the time of writing this script) there's no way to do a GET call to read all the videos posted to a specific Facebook page in order to determine which videos have been uploaded for the purposes of deduplication.